// libs/sales/customer-registration/data-access/src/lib/entities/interfaces/address-store.interface.ts
export interface AddressStoreInterface {
  addressType: string | undefined;
  addressId?: string;
  streetId?: string;
  tinaLocationId?: string;
  size?: number;
  country?: string;
  boxNumber?: string;
  subHouseNumber?: string;
  houseNumber?: string;
  postalCode?: string;
  municipality?: string;
  street?: string;
}